07-strider-goose-offroad-terrain/requirements.txt
openai>=1.50
python-dotenv>=1.0
anthropic>=0.39
gradio>=4.0
Pillow>=10.0
numpy>=1.26
